integrated yolo models and made changed to front end

+++ backend/app.py
@@ -1,49 +1,44 @@
-import os
 import json
+import os
+import re
+os.environ["KMP_DUPLICATE_LIB_OK"] = "TRUE"
+import cv2
+import matplotlib
+matplotlib.use('Agg')
+import google.generativeai as genai # type: ignore
 import time
-import re 
-from flask import Flask, request, jsonify # pyright: ignore[reportMissingImports]
-import yt_dlp # pyright: ignore[reportMissingModuleSource]
-import google.generativeai as genai  # pyright: ignore[reportMissingImports]
-from dotenv import load_dotenv # pyright: ignore[reportMissingImports]
-from flask_cors import CORS # pyright: ignore[reportMissingModuleSource]
-
+import yt_dlp  # type: ignore
+from dotenv import load_dotenv 
+from flask import Flask, request, jsonify  # type: ignore
+from flask_cors import CORS  # type: ignore
+from ultralytics import YOLO # type: ignore
 load_dotenv()
 app = Flask(__name__)
 CORS(app)
+classifier_model = YOLO('https://huggingface.co/francm2/image-classification-model/resolve/main/yolov8n-classification-model.pt') 
+detection_model = YOLO('https://huggingface.co/Ultralytics/YOLOv8/resolve/main/yolov8n.pt')
 
-
-try:
-    genai.configure(api_key=os.environ["GEMINI_API_KEY"])
-except KeyError:
-    print("Error: GEMINI_API_KEY not found. Make sure it's set in your .env file.")
-    exit()
 
 
 ANALYSIS_PROMPT = """
 You are an expert AI video analysis system. Your goal is to detect
 AI-generated content (deepfakes, generative video).
-Analyze the provided video and return *only* a valid JSON object
+Analyze the provided image and return *only* a valid JSON object
 in the following format:
 {
   "isAI": boolean,
   "confidence": number (0-100),
   "visualArtifacts": number (0-100),
-  "audioAnomalies": number (0-100),
-  "motionPatterns": number (0-100),
-  "faceAnalysis": number (0-100)
+  "faceAnalysis": number (0-100),
   "explanation": "Your detailed, one-paragraph analysis. Explain *why*
-  you made this decision. Be specific about what you see or hear."
+  you made this decision. Be specific about what you see."
 }
 """
-
-
 generation_config = {
     "response_mime_type": "application/json",
 }
 
-
-model = genai.GenerativeModel(
+gemini_model = genai.GenerativeModel(
     'models/gemini-flash-latest',
     generation_config=generation_config
 )
@@ -56,20 +51,61 @@ def extract_json_from_text(text):
         return match.group(0)
     return None
 
+def gemini_analysis():
+    frame_path = "suspicious_frame.jpg" 
+    try:
+        genai.configure(api_key=os.environ["GEMINI_API_KEY"])
+    except KeyError:
+        print("Error: GEMINI_API_KEY not found.")
+        return {"error": "Server configuration error"}, 500
+
+    print(f"Uploading {frame_path} to Gemini...")
+    frame_file = genai.upload_file(path=frame_path)
+    while frame_file.state.name == "PROCESSING":
+        time.sleep(1) 
+        frame_file = genai.get_file(frame_file.name)
+
+    if frame_file.state.name == "FAILED":
+        return {"error": "Google failed to process the image."}, 500
+
+    print(f"File uploaded: {frame_file.name}") 
+
+    try:
+
+        print("Analyzing with Gemini...")
+        response = gemini_model.generate_content([ANALYSIS_PROMPT, frame_file])
+        print(f"Raw JSON response: {response.text}")
+        try:
+            result_data = json.loads(response.text)
+        except json.JSONDecodeError:
+            print("Response not clean, attempting extraction...")
+            clean_json = response.text.replace('```json', '').replace('```', '').strip()
+            result_data = json.loads(clean_json)
+
+        return result_data 
+
+    except Exception as e:
+        print(f"Analysis failed: {e}")
+        return {"error": str(e)}
+
+    finally:
+        if frame_file:
+            print(f"Deleting cloud file: {frame_file.name}")
+            genai.delete_file(frame_file.name)
+
 
 @app.route('/analyze', methods=['POST'])
 def analyze_video():
     data = request.json
     if data is None:
-        return jsonify({"error": "No JSON data. Did you forget 'Content-Type: application/json'?"}), 400
+        return jsonify({"error": "No JSON data"}), 400
     if 'url' not in data:
-        return jsonify({"error": "No 'url' provided in JSON"}), 400
+        return jsonify({"error": "No 'url' provided"}), 400
 
     video_url = data['url']
     downloaded_filename = ""
 
     try:
-      
         print(f"Downloading video from: {video_url}")
         ydl_opts = {
             'outtmpl': 'temp_video.%(ext)s',
@@ -77,61 +113,60 @@ def analyze_video():
             'quiet': True,
         }
 
-        with yt_dlp.YoutubeDL(ydl_opts) as ydl: # pyright: ignore[reportArgumentType]
+        with yt_dlp.YoutubeDL(ydl_opts) as ydl: # type: ignore
             info = ydl.extract_info(video_url, download=True)
             downloaded_filename = ydl.prepare_filename(info)
-
+        
         print(f"Video downloaded: {downloaded_filename}")
+        results = classifier_model(downloaded_filename, stream=True)
+        ai_score_sum = 0
+        frame_count = 0
+        max_ai_score = -1.0
+        best_frame_img = None
+        for r in results:
+            ai_prob = float(r.probs.data[1]) 
+            ai_score_sum += ai_prob
+            frame_count += 1
+            if(ai_prob>max_ai_score):
+                max_ai_score = ai_prob
+                best_frame_img = r.orig_img.copy()
 
-    
-        print("Uploading file to Gemini...")
-        video_file = genai.upload_file(path=downloaded_filename)
+        avg_score = ai_score_sum / frame_count if frame_count >0 else 0
+
+        result = {}
+        result["url"] = video_url
+        result["timestamp"] = int(time.time()*1000)
+        print(f"Found suspicious frame with AI Score: {round(max_ai_score, 2)}")
+        if best_frame_img is not None and max_ai_score>0.1:
+            print(f"Found suspicious frame with AI Score: {round(max_ai_score, 2)}")
+            cv2.imwrite("suspicious_frame.jpg",best_frame_img)
+            print("Sending to Gemini")
+            result = gemini_analysis()
+            if result.get("isAI") is False :
+                result["explanation"] = "Our propietary detection system did not identify any frames that can potentially be deemed as AI generated."
         
-        while video_file.state.name == "PROCESSING":
-            print("Waiting for Google to process the video...")
-            time.sleep(5)
-            video_file = genai.get_file(video_file.name)
+            return jsonify(result), 200
+        else:
+            print("No frames sent for extra detection")
+            result["isAI"] = "false"
+            result["explanation"] = "Our propietary detection system did not identify any frames that can potentially be deemed as AI generated."
+            result["confidence"] = 100-round(avg_score*100,2)
+            return jsonify(result), 200
         
-        if video_file.state.name == "FAILED":
-            return jsonify({"error": "Google failed to process the video file."}), 500
-
-        print(f"File uploaded successfully: {video_file.name}")
-
-     
-        print("Analyzing video with Gemini...")
-        response = model.generate_content([ANALYSIS_PROMPT, video_file])
-        genai.delete_file(video_file.name)
-        print("Removed file from Google's servers.")
-
-    
-        print(f"Raw JSON response: {response.text}")
-        
-        try:
-            result_json = json.loads(response.text)
-        except json.JSONDecodeError:
-            print("Response was not clean JSON, trying to extract...")
-            json_string = extract_json_from_text(response.text)
-            if not json_string:
-                return jsonify({"error": "Failed to parse AI response. Check backend logs."}), 500
-            result_json = json.loads(json_string)
-
-        result_json["url"] = video_url
-        result_json["timestamp"] = time.time()
-        
-        print(f"Analysis complete: {result_json['isAI']} ({result_json['confidence']}%)")
-        return jsonify(result_json)
 
     except Exception as e:
         print(f"An error occurred: {e}")
         return jsonify({"error": str(e)}), 500
     
     finally:
-   
-        if os.path.exists(downloaded_filename):
+        if downloaded_filename and os.path.exists(downloaded_filename):
             os.remove(downloaded_filename)
             print(f"Removed temp file: {downloaded_filename}")
-
+        if os.path.exists("suspicious_frame.jpg"):
+            os.remove("suspicious_frame.jpg")
+            
+            print(f"Removed file")
 
 if __name__ == "__main__":
-    app.run(host='0.0.0.0', port=5001, debug=True)
 
+    app.run(host='0.0.0.0', port=5002, debug=True, use_reloader=False)
